Fix documentation for Sobel Filter
The sobel filter specifies that "scale" is a parameter, when it is not.
Remove the line conerning "scale"

"dx" and "dy" specify that the derivative can be 1 to 3
Change the documentation to specify the derivative can be from 0 to 3
This is as according to the example usage on the sobel filter documentation.