Fixed bug where newmask as destroyed in debug mode
In debug mode, cv2.findContours is run on the newmask image to detect
contours so that they can be drawn on the original image for debug
purposes. This alters newmask, which is then returned by the function.
np.copy was added to input a copy of newmask to findContours so that
newmask remains intact.