Added detailed guide to updating to v3.0dev2
The section Updating to v3 has a detailed description of the changes to every function call. As we make additional changes in the v3.0dev2 pull request we can update this section so that we keep track of everything that changed and give people a clear guide of how to update their existing code after the release.