detect_peaks_trim
Commented out parts of the _detect_peaks function which is used within the triangle function of the thresholding methods subpackage. There were many if statements that weren't covered and this is because the only place _detect_peaks is used is within the triangle function and all of the parameters are hardcoded, so many if statements will never be true.