crop_position_mask_fxn_edit
After trying to cover if statements I traced back maskv to find that mask is spliced down earlier in the function if it has 3 dimensions. So if a color mask is fed in, it will be spliced down and will never satisfy len(np.shape(maskv))==3 the rest of the function.