edit analyze_bound_horizontal
Changed the analyze_bound_horizontal function so that the line position variable describes from the top down rather than from the bottom up like it had been working. The rest of the functions work this way (top left is the starting point) so this will make the function consistent with others. Also updated unit test and documentation.