update ori_img, mask, kept_cnt, & kept_hierarchy
w_back image was keeping only the largest contour (and child contours within that contour) but ori_img and mask were plotting as if roi_type was set to 'partial'. The same with kept_cnt and kept_hierarchy (tested with object composition downstream) so I updated those variables that get returned.