Remove bins input
The function would have only accepted 8-bit images for other reasons, as-is, so the maximum number of bins was 256. The bins input would let you scale down the number of bins if desired. However, this can be done downstream and due to other upcoming changes with how the output values are reported, bins would have become a complicated value. This simplifies the function for everyone