update cluster_contours
pcv.cluster_contours would error out if nrow or ncol > 9 since the row and col number were extracted assuming item at index 0 was col, item at index 1 was ",", and item at index 2 was row. Instead, split on the comma in the case where nrow or ncol is greater than or equal to 10.