update hyperspectral functions to include pseudo_rgb
We want to attach a pseudo_rgb image object to a spectral_data class instance if the datatype is a datacube. This way users can access the pseudo_rgb image if they want to save it or otherwise access it directly.  If the datatype is not a hyperspectral datacube then just pass none.