delete code that isn't doing anything
Originally we were trying to requiring timestamp format for --adaptor=filename but that is only true if --dates is provided. also, since timestampformat has a default value  I don't know how to check if a user provided format is provided  - if identical to the default how do we know if it is/isn't supposed to be that way?   So i just deleted these things.