Update extract_index.py
Add new feature to extract_index function such that the indices for pixels having NAN and/or Inf would be returned as the second output of the function.
Add new feature such that the theoretical range for the VI would be printed out; a warning message of the existance of NAN and/or Inf in the output will also be printed if that's the case.
Fix typo/wrong formula in extract_index.